+++ controllers/ListUsersController.php
@@ -9,18 +9,21 @@ class ListUsersController
     function __construct($get, $post, $route) {
         $this->dao = new UserDAO();
         $this->view = new ListUserPageView();
+        if(SecurityHelper::isAdmin()){
 //        var_dump('route', $route, 'post', $post);
-        if($get && isset($get['delete'])) {
-            $this->delete($get['delete']);
+            if($get && isset($get['delete'])) {
+                $this->delete($get['delete']);
+            }
+            $this->data=$this->dao->fetchAll();
+            $this->displayOne($this->data);
+        }else{
+            header('Location: /MyBuilding/');
         }
-        $this->data=$this->dao->fetchAll();
-        $this->displayOne($this->data);
     }
 
     function displayOne($data) {
         echo $this->view->displayOne($data);
     }
-
 
     function delete($id) {
         $this->dao->delete($id);
